Made Borrower ID column correctly print -1

+++ Final Project Code/src/inventoryitem.cpp
@@ -78,7 +78,7 @@ Row InventoryItem::provideRow() const {
         name,
         author,
         publisher,
-        std::to_string(borrowerID)
+        std::to_string(static_cast<int64_t>(borrowerID))
     });
 }
